// rent-acar-angular/src/app/models/korisnik.model.ts
export interface KorisnikD {
    email: string;
    sifra: string;
    ime: string;
    prezime: string;
    jmbg: string;
    drzava: string;
    grad: string;
    adresa: string;
    telefon: string;
    broj_pasosa: string;
    broj_vozacke: string;
}

export interface Korisnik {
    id: number;
    email: string;
    sifra: string;
    ime: string;
    prezime: string;
    jmbg: string;
    drzava: string;
    grad: string;
    adresa: string;
    telefon: string;
    broj_pasosa: string;
    broj_vozacke: string;
}